+++ src/Listener.php
@@ -3,15 +3,16 @@
 /**
  * Title: OmniKassa listener
  * Description:
- * Copyright: Copyright (c) 2005 - 2015
+ * Copyright: Copyright (c) 2005 - 2016
  * Company: Pronamic
+ *
  * @author Remco Tolsma
  * @version 1.0.0
  */
 class Pronamic_WP_Pay_Gateways_OmniKassa_Listener implements Pronamic_Pay_Gateways_ListenerInterface {
 	public static function listen() {
-		if ( 
-			filter_has_var( INPUT_POST, 'Data' ) 
+		if (
+			filter_has_var( INPUT_POST, 'Data' )
 				&&
 			filter_has_var( INPUT_POST, 'Seal' )
 		) {